refactor: Move the file inside the attribute-options block
