+++ plugins/woocommerce/client/blocks/assets/js/blocks/add-to-cart-with-options/variation-selector/attribute-options/frontend.ts
@@ -7,7 +7,7 @@ import { store, getContext, getElement } from '@wordpress/interactivity';
 /**
  * Internal dependencies
  */
-import setStyles from '../../set-styles';
+import setStyles from './set-styles';
 
 type Option = {
 	value: string;
